+++ tests/test_clear_approval_mail.py
@@ -124,7 +124,7 @@ def test_clear_approval_mail_tolerates_unreadable_file(bridget):
 
 def _fake_mg_idea(args):
     if args[:1] == ['show']:
-        return 0, 'ID: mg-aaaa\nType: idea\nStatus: available\n', ''
+        return 0, 'ID: mg-aaaa\nType: idea\nAssignee: architect\nStatus: available\n', ''
     if args[:2] == ['mail', 'send']:
         return 0, '', ''
     if args[:1] == ['unshelve']: